pyproject.toml
[tool.poetry]
name = "bot"
version = "2.0.0"
description = "The official StudSec bot written in Python"
license = "GPL-3.0-only"
authors = [
    "Robert Klink <roberthklink@gmail.com>",
    "Aidan <aidands@protonmail.com>"
]

[tool.poetry.dependencies]
python = "^3.8"
"discord.py" = "^2.3.2"
python-dotenv = "^1.0.1"
termcolor = "^2.4.0"
selenium = "^4.17.2"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pylint = "^3.0.3"

[tool.poetry.scripts]
bot = "bot.main:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
